(v0.08) Improve interrupt handling & Bug fixes
Major refactoring of interrupt and PIC handling:

- Fixed IRQ stub naming and vector handling (use actual vector numbers, not offsets)
- Made outb/inb functions explicitly unsafe and added safety checks
- Improved PIC masking logic with dedicated mask/unmask functions
- Fixed zero-division and overflow issues in PIT sleep/uptime functions
- Added frequency validation in PIT
- Refactored GDT initialization to use constants instead of magic numbers
- Improved interrupt safety with proper disable/enable guards
- Fixed register preservation in VGA pixel operations
- Added special case handling for 0 in decimal conversion
- Various assembly-level fixes and stack alignment improvements

+++ source/kernel32/src/x86/memory.rs
@@ -16,6 +16,6 @@ pub struct E820Entry {
 #[repr(C, packed)]
 #[derive(Copy, Clone)]
 pub struct E820Map {
-    pub entry_count: u32,
+    pub entry_count: u16,
     pub map: [E820Entry; 64],
 }
